--- src/domain/repositories/DiseaseOutbreakEventRepository.ts
@@ -1,6 +1,0 @@
-import { FutureData } from "../../data/api-futures";
-import { ConfigLabel, Id } from "../entities/Ref";
-
-export interface DiseaseOutbreakEventRepository {
-    getConfigStrings(): FutureData<ConfigLabel[]>;
-}
